extract manager to its how file
