+++ src/server/mod.rs
@@ -7,6 +7,7 @@ mod default;
 pub mod error;
 mod inter;
 mod listener;
+mod manager;
 mod monitor;
 mod nix_utils;
 mod relaunch;
